add Security to AppPreferences
